Add basic SendFile implementation
Use the GET_CONTENT intent to open an external file selector dialog.
In java land we copy the file into our apps local cache so we can
handle it as just a normal file in our Go code. This is necessary as
the GET_CONTENT intent doesn't proved access to the file directly but
gives you an open file stream you can work with.