Fix minsdk version to something > 0
We now explicitly set minsdk when calling gogio to build our aar file.
Previously when we were building an apk the tool was setting the
minversion to 16 (I think). When you build an aar it will set the
minversion to 0 if you don't give it a min version.

Remove old jgo.jar references from the makefile.