Support android share intent
This adds initial support for android the share intent. Instead of
opening the app and choosing a file, you can share the file from a
different app.

The implementation is a little awkward. Normally you would receive
this intent in the main application activity. Gio owns that so we
can't do that in this case. So instead we make a new activity that
receives the event and then switches to the main activity.

Currently nothing prevents you from sharing from the UI and from a
share intent concurrently. The UI doesn't support this at all so the
first action will just disappear (although the messages will probably
get messed up).

We should add proper state management for this.

Fixes: #2